+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-* Fixed: Fix the module schema test if the module has no helper files
+* Fixed: Fix the module schema test if the module has no helper files or a single helper file
 
 ## 3.4.0 - 2026-06-18